JWT claims and authoriztion

+++ resources/item.py
@@ -5,7 +5,7 @@ from flask_smorest import Blueprint, abort
 from schemas import ItemSchema,ItemUpdateSchema
 
 #Predicting our endpoints, If a user doesn't have the correct JWT token  
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required,get_jwt
 
 from sqlalchemy.exc import SQLAlchemyError
 
@@ -25,6 +25,10 @@ class Item(MethodView):
     
     @jwt_required()
     def delete(self, item_id):
+        jwt=get_jwt()  # use of claim
+        if not jwt.get("is_admin"):
+            abort(401,message="Admin privilage required.")
+
         item=ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
